refactor: extract shared price-sources module
- Create /src/lib/services/price-sources.ts with shared logic
- Extract fetchYahooPrice, fetchCoinGeckoPrice, and fetchPriceWithRetry
- Create PriceCache class with LRU eviction
- Add proper TypeScript interfaces (PriceMetadata, PriceResult, etc.)
- Refactor /src/app/api/prices/[symbol]/route.ts to use shared module
- Refactor /src/app/api/prices/batch/route.ts to use shared module
- Saves ~478 lines of duplicated code (220 + 258)

Phase 2, Tasks 2.1-2.2 complete
